Refactor API structure by introducing configuration and client modules. Implement title generation and token counting services. Update README for development server instructions.
